fix: jset test WIP
